Move clue functions to the clue_selection module
